autopkgtests: Incorporate (and fix) the existing testsuite, and add further tests
Previously there was a single autopkgtest plus 4 tests run via python in tests/
This change makes all of these pass, and runs them all from autopkgtests.

It drops the python test.py, which did not work with python3, but keeps the substance.
It also adds tests of the various REPORTLEVEL and of some simple "error conditions"