logcheck.{postinst,postrm}: Update to only change permissions on /etc/logcheck
Before, logcheck was changing permissions of all files in /etc/logcheck on install,
including rules put there by other packages, and not restoring them on purge. This
is, rightly, flagged by piuparts as an issue.

This change simplifies the change in permissons to just the /etc/logcheck directory,
and on purge restores it to root:root ownership. It also restores the contents, but
once this version is in stable that could be further simplified to avoid a recursive
use of chown -R, which is safe here (given it changes to root) but flagged by lintian.

It also moves header.txt to /etc so it can be a conffile